refactor: 상품목록 카드 컴포넌트로 분리
