+++ src/pages/index.tsx
@@ -1,9 +1,6 @@
-const Home = () => {
-  {
-    /* [TODO] json-server에서 데이터 가져오는 것으로 수정 */
-  }
-  const TEMP_ARRAY = Array(16).fill(null);
+import Card from '@/components/Card';
 
+const Home = () => {
   return (
     <div className="flex flex-col items-center justify-center">
       <header>
@@ -12,22 +9,7 @@ const Home = () => {
         </div>
       </header>
 
-      <body>
-        <div className="grid grid-cols-4 gap-8">
-          {TEMP_ARRAY.map((_, index) => {
-            return (
-              <div
-                key={index}
-                className="flex transform flex-col rounded-lg p-4 shadow-sm transition-transform hover:scale-105 hover:cursor-pointer hover:shadow-2xl"
-              >
-                <div className="h-60 w-60 bg-gray-200"></div>
-                <div className="text-center text-lg">상품 이름</div>
-                <div className="text-center text-sm text-gray-500">가격</div>
-              </div>
-            );
-          })}
-        </div>
-      </body>
+      <Card />
     </div>
   );
 };
